test(engine): merge predicted-gate-engine-branch-coverage.test.ts into predicted-gate-engine.test.ts (#8575)
predicted-gate-engine-coverage.test.ts (1432 lines) and
predicted-gate-engine-branch-coverage.test.ts (708 lines) were two
disconnected #2283 bolt-ons against the same gate-advisory/
predicted-gate-engine module family. Merge into one file
(predicted-gate-engine.test.ts) as two describe blocks, keeping every
case verbatim -- a pure relocation, no case retitled or dropped.

Renamed the smaller file's local REPO const to BRANCH_REPO to resolve
the one identifier collision between the two files (its shape differs
from the larger file's REPO -- a shared literal object -- so they
cannot be merged into a single fixture without changing what several
tests exercise). No other identifiers collided.

Verified zero coverage regression across every module the merged file
exercises (gate-advisory.ts, predicted-gate-engine.ts, cla-check.ts,
pre-merge-checks.ts, change-guardrail.ts, duplicate-winner.ts) --
before/after branch and line coverage identical on all six.
